feat: Api 요청 파일 분리 및 에러 처리 추가
커밋 목록을 요청하는 Api 요청 파일을 하위 폴더에 별개의 파일로 분리 하였습니다.
api요청이 이루어지지 않았을 때의 에러처리도 추가 하였습니다.
- UI 작업시에 해당 에러 처리를 이용할 수 있게 형식적으로 작성하였습니다.
